+++ Tests/Class.test.ps1
@@ -1,5 +1,6 @@
 
 Import-Module Ldbc
+. .\Class.lib.ps1
 
 function ClassExample {
 
@@ -72,5 +73,38 @@ task TODO {
 		catch {
 			assert ("$_" -like '*Object reference not set to an instance of an object.*')
 		}
+	}
+}
+
+task DbRef {
+	$customer = [Customer]@{CustomerId = 1; Name = 'John'}
+	$product1 = [Product]@{ProductId = 1; Name = 'Foo'}
+	$product2 = [Product]@{ProductId = 2; Name = 'Bar'}
+	$order = [Order]@{OrderId = 1; Customer = $customer; Products = $product1, $product2}
+
+	Use-LiteDatabase -Script {
+		$customers = Get-LiteCollection customers
+		$products = Get-LiteCollection products
+		$orders = Get-LiteCollection orders
+
+		$order.Products | Add-LiteData $products
+		Add-LiteData $customers $customer
+		Add-LiteData $orders $order
+
+		# raw, no refs
+		$r = Get-LiteData $orders
+		equals "$r" '{"_id":1,"Customer":{"$id":1,"$ref":"customers"},"Products":[{"$id":1,"$ref":"products"},{"$id":2,"$ref":"products"}]}'
+
+		# raw, with refs
+		$r = Get-LiteData $orders -Include '$.Customer', '$.Products[*]'
+		equals "$r" '{"_id":1,"Customer":{"_id":1,"Name":"John"},"Products":[{"_id":1,"Name":"Foo"},{"_id":2,"Name":"Bar"}]}'
+
+		# typed, no refs
+		$r = Get-LiteData $orders -As Order
+		equals "$r" '#1 #1  [#1 , #2 ]'
+
+		# typed, with refs
+		$r = Get-LiteData $orders -As Order -Include '$.Customer', '$.Products[*]'
+		equals "$r" '#1 #1 John [#1 Foo, #2 Bar]'
 	}
 }